Fix circular dependencies in `navigator` package
